requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Django
django-cors-headers==4.3.1
django-extensions==3.2.3
django-sslserver==0.22
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
idna==3.7
mssql-django
multidict==6.0.5
mysqlclient==2.2.4
pillow==10.2.0
pyaes==1.6.1
pyasn1==0.6.0
pycparser==2.22
PyJWT==2.8.0
pyodbc==5.1.0
pyOpenSSL==24.1.0
python-dotenv==1.0.1
pythonansi==1.0.2
pytz==2024.1
requests==2.32.2
rsa==4.9
sqlparse==0.5.0
Telethon==1.35.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
yarl==1.9.4
